vswitch: move most of the implementation to the cpp file
We don't need most of vswitch in the places where it is included.
